refactor: split Codex hook setup tests
